+++ testing/simulated/execution/geth.go
@@ -20,7 +20,7 @@
 // MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE, NON-INFRINGEMENT, AND
 // TITLE.
 
-package simulated
+package execution
 
 import (
 	"fmt"
